+++ lib/src/ui.dart
@@ -1,10 +1,11 @@
 import 'dart:typed_data';
 import 'dart:ui' as ui;
+
 import 'main.dart';
 
 extension TagLibUtils on TagLib {
   Future<ui.Image?> getArtwork(String filename) async {
-    Uint8List? imgBytes = getArtworkBytes(filename);
+    Uint8List? imgBytes = await getArtworkBytes(filename);
     if (imgBytes == null) return null;
     ui.Codec codec = await ui.instantiateImageCodec(imgBytes);
     ui.FrameInfo frame = await codec.getNextFrame();