+++ lib/src/utils.dart
@@ -3,18 +3,25 @@ import 'dart:io';
 
 const String _libName = 'taglib_ffi';
 DynamicLibrary loadLibrary(String path, bool test) {
+  String fullpath = '';
+
   if (Platform.isMacOS || Platform.isIOS) {
     if (test) {
-      return DynamicLibrary.open('$path/lib$_libName.dylib');
+      fullpath = '$path/lib$_libName.dylib';
     } else {
-      return DynamicLibrary.open('$_libName.framework/$_libName');
+      fullpath = '$_libName.framework/$_libName';
     }
+  } else if (Platform.isAndroid || Platform.isLinux) {
+    fullpath = '$path/lib$_libName.so';
+  } else if (Platform.isWindows) {
+    fullpath = '$path/$_libName.dll';
+  } else {
+    throw UnsupportedError('Unknown platform: ${Platform.operatingSystem}');
   }
-  if (Platform.isAndroid || Platform.isLinux) {
-    return DynamicLibrary.open('$path/lib$_libName.so');
+
+  if (File(fullpath).existsSync()) {
+    return DynamicLibrary.open(fullpath);
+  } else {
+    return DynamicLibrary.executable();
   }
-  if (Platform.isWindows) {
-    return DynamicLibrary.open('$path/$_libName.dll');
-  }
-  throw UnsupportedError('Unknown platform: ${Platform.operatingSystem}');
 }